+++ index.js
@@ -17,6 +17,7 @@ var React = require("react");
 var PropTypes = require("prop-types");
 var invariant = require("invariant");
 var deepClone = require("@jsbits/deep-clone");
+var UNDEF = undefined;
 /**
  * We are running in a Mac-like OS?
  */
@@ -103,7 +104,7 @@ exports.setDefaults = function (defaults) {
     var defs = assign(_Conf.defs, deepClone(defaults));
     // check and format class names, if any
     if (defs.className) {
-        defs.className = classAsStr(defs.className).trim() || undefined;
+        defs.className = classAsStr(defs.className).trim() || UNDEF;
     }
     // cleanup empty properties
     pack(defs);
@@ -120,6 +121,10 @@ exports.setSizes = function (sizes) {
     invariant(isObject(sizes), 'The sizes must be an object.');
     pack(assign(_Conf.sizes, sizes));
 };
+var expandFrom = {
+    color: ['fill', 'stroke'],
+    size: ['width', 'height'],
+};
 /**
  * Renders a SVG Ionicon
  */
@@ -128,6 +133,15 @@ var IonIcon = /** @class */ (function (_super) {
     function IonIcon() {
         return _super !== null && _super.apply(this, arguments) || this;
     }
+    IonIcon.prototype.expandAttr = function (opts, name, value) {
+        var _a = expandFrom[name], p1 = _a[0], p2 = _a[1];
+        if (opts[p1] !== UNDEF) {
+            opts[p1] = value;
+        }
+        if (opts[p2] !== UNDEF) {
+            opts[p2] = value;
+        }
+    };
     /**
      * Merge the user properties with the defaults, taking care to preserve
      * the classes and styles of both.
@@ -135,6 +149,15 @@ var IonIcon = /** @class */ (function (_super) {
     IonIcon.prototype.mergeDefs = function (props) {
         var defs = _Conf.defs;
         var keys = keyArray(defs);
+        // Example of precedence [fill, stroke] = troke:
+        // fill = user.fill -> defs.fill -> user.color -> defs.color
+        // even if any of them is `null`, it will overrite.
+        if (props.color !== UNDEF) {
+            this.expandAttr(props, 'color', props.color);
+        }
+        if (props.size !== UNDEF) {
+            this.expandAttr(props, 'size', props.size);
+        }
         for (var i = 0; i < keys.length; i++) {
             var k = keys[i];
             switch (k) {
@@ -147,6 +170,10 @@ var IonIcon = /** @class */ (function (_super) {
                     props.style = props.style
                         ? assign(assign({}, props.style), defs.style)
                         : defs.style;
+                    break;
+                case 'color':
+                case 'size':
+                    this.expandAttr(props, k, defs[k]);
                     break;
                 default:
                     if (!props.hasOwnProperty(k)) {
@@ -169,16 +196,6 @@ var IonIcon = /** @class */ (function (_super) {
         }
         // name & innerRef are out, merge with defaults before color & size
         this.mergeDefs(opts);
-        var color = opts.color;
-        if (color != null) {
-            delete opts.color;
-            opts.fill = opts.stroke = color;
-        }
-        var size = opts.size;
-        if (size != null) {
-            delete opts.size;
-            opts.width = opts.height = typeof size === 'string' && _Conf.sizes[size] || size;
-        }
         // Guess whether the "iOS" style should be used with double-style icons.
         var ios = isMacLike;
         if (opts.mode) {